# Conflicts:
#	src/Internal/Data/Queue.h

+++ src/Internal/Algorithms/Math.h
@@ -3,27 +3,50 @@
 
 namespace Emblate
 {
+    /**
+     * Computes the absolute value of the element.
+     *
+     * @tparam T The type of the elements.
+     * @param val Value to compute the absolute value.
+     * @return Absolute value of the element.
+     */
     template<typename T>
     T abs(T val)
     {
         return val >= 0 ? val : -val;
     }
 
+    /**
+     * Computes the largest integer value not greater than _val_.
+     *
+     * @tparam T The type of the elements.
+     * @param val Value to compute the absolute value.
+     * @return Largest integer value not greater than _val_.
+     */
+    // TODO: Return type must be equal to T.
     template<typename T>
-    inline int floor(T val)
+    inline size_t floor(T val)
     {
-        return (int)val;
+        return static_cast<size_t>(val);
     }
 
+    /**
+     * Computes the least integer value not less than _val_.
+     *
+     * @tparam T The type of the elements.
+     * @param val Value to compute the absolute value.
+     * @return Least integer value not less than _val_.
+     */
+    // TODO: Return type must be equal to T.
     template<typename T>
-    inline int ceil(T val)
+    inline size_t ceil(T val)
     {
-        if (val >= 0 && val - (int)val > 0)
+        if ((val >= 0) && (val - static_cast<size_t>(val) > 0))
         {
-            return (int)(val + 1);
+            return static_cast<size_t>(val + 1);
         }
 
-        return (int)(val);
+        return static_cast<size_t>(val);
     }
 
     template<typename T>
